+++ src/ggrc/models/comment.py
@@ -15,7 +15,10 @@ from werkzeug.exceptions import BadRequest
 
 from ggrc import builder
 from ggrc import db
-from ggrc.models.custom_attribute_definition import CustomAttributeDefinition
+from ggrc.models.custom_attribute_definition import (
+    CustomAttributeDefinition,
+    CustomAttributeDefinitionFK,
+)
 from ggrc.models.deferred import deferred
 from ggrc.models.revision import Revision
 from ggrc.models.mixins import base, synchronizable
@@ -198,8 +201,17 @@ def get_objects_to_reindex(obj):
   return list(itertools.chain(*queries))
 
 
-class Comment(Roleable, Relatable, Described, Notifiable,
-              base.ContextRBAC, Base, Indexed, db.Model):
+class Comment(
+    CustomAttributeDefinitionFK,
+    Roleable,
+    Relatable,
+    Described,
+    Notifiable,
+    base.ContextRBAC,
+    Base,
+    Indexed,
+    db.Model,
+):
   """Basic comment model."""
   __tablename__ = "comments"
 
@@ -211,15 +223,6 @@ class Comment(Roleable, Relatable, Described, Notifiable,
   ), 'Comment')
   revision = db.relationship(
       'Revision',
-      uselist=False,
-  )
-  custom_attribute_definition_id = deferred(db.Column(
-      db.Integer,
-      db.ForeignKey('custom_attribute_definitions.id', ondelete='SET NULL'),
-      nullable=True,
-  ), 'Comment')
-  custom_attribute_definition = db.relationship(
-      'CustomAttributeDefinition',
       uselist=False,
   )
 
@@ -234,14 +237,11 @@ class Comment(Roleable, Relatable, Described, Notifiable,
   # REST properties
   _api_attrs = reflection.ApiAttributes(
       "assignee_type",
-      reflection.Attribute("custom_attribute_revision",
-                           create=False,
-                           update=False),
-      reflection.Attribute("custom_attribute_revision_upd",
-                           read=False),
-      reflection.Attribute("header_url_link",
-                           create=False,
-                           update=False),
+      reflection.Attribute(
+          "header_url_link",
+          create=False,
+          update=False,
+      ),
   )
 
   _sanitize_html = [
@@ -253,10 +253,6 @@ class Comment(Roleable, Relatable, Described, Notifiable,
       ReindexRule("Relationship", reindex_by_relationship),
   ]
 
-  _aliases = {
-      "custom_attribute_definition": "custom_attribute_definition",
-  }
-
   @builder.simple_property
   def header_url_link(self):
     """Return header url link to comment if that comment related to proposal
@@ -267,86 +263,6 @@ class Comment(Roleable, Relatable, Described, Notifiable,
     if self.initiator_instance.status == proposed_status:
       return "proposal_link"
     return ""
-
-  @classmethod
-  def eager_query(cls, **kwargs):
-    query = super(Comment, cls).eager_query(**kwargs)
-    return query.options(
-        orm.joinedload('revision'),
-        orm.joinedload('custom_attribute_definition')
-           .undefer_group('CustomAttributeDefinition_complete'),
-    )
-
-  def log_json(self):
-    """Log custom attribute revisions."""
-    res = super(Comment, self).log_json()
-    res["custom_attribute_revision"] = self.custom_attribute_revision
-    return res
-
-  @builder.simple_property
-  def custom_attribute_revision(self):
-    """Get the historical value of the relevant CA value."""
-    if not self.revision:
-      return None
-    revision = self.revision.content
-    cav_stored_value = revision['attribute_value']
-    cad = self.custom_attribute_definition
-    return {
-        'custom_attribute': {
-            'id': cad.id if cad else None,
-            'title': cad.title if cad else 'DELETED DEFINITION',
-        },
-        'custom_attribute_stored_value': cav_stored_value,
-    }
-
-  def custom_attribute_revision_upd(self, value):
-    """Create a Comment-CA mapping with current CA value stored."""
-    ca_revision_dict = value.get('custom_attribute_revision_upd')
-    if not ca_revision_dict:
-      return
-    ca_val_dict = self._get_ca_value(ca_revision_dict)
-
-    ca_val_id = ca_val_dict['id']
-    ca_val_revision = Revision.query.filter_by(
-        resource_type='CustomAttributeValue',
-        resource_id=ca_val_id,
-    ).order_by(
-        Revision.created_at.desc(),
-    ).limit(1).first()
-    if not ca_val_revision:
-      raise BadRequest("No Revision found for CA value with id provided under "
-                       "'custom_attribute_value': {}"
-                       .format(ca_val_dict))
-
-    self.revision_id = ca_val_revision.id
-    self.revision = ca_val_revision
-
-    # Here *attribute*_id is assigned to *definition*_id, strange but,
-    # as you can see in src/ggrc/models/custom_attribute_value.py
-    # custom_attribute_id is link to custom_attribute_definitions.id
-    # possible best way is use definition id from request:
-    # ca_revision_dict["custom_attribute_definition"]["id"]
-    # but needs to be checked that is always exist in request
-    self.custom_attribute_definition_id = ca_val_revision.content.get(
-        'custom_attribute_id',
-    )
-
-    self.custom_attribute_definition = CustomAttributeDefinition.query.get(
-        self.custom_attribute_definition_id,
-    )
-
-  @staticmethod
-  def _get_ca_value(ca_revision_dict):
-    """Get CA value dict from json and do a basic validation."""
-    ca_val_dict = ca_revision_dict.get('custom_attribute_value')
-    if not ca_val_dict:
-      raise ValueError("CA value expected under "
-                       "'custom_attribute_value': {}"
-                       .format(ca_revision_dict))
-    if not ca_val_dict.get('id'):
-      raise ValueError("CA value id expected under 'id': {}"
-                       .format(ca_val_dict))
-    return ca_val_dict
 
 
 class ExternalComment(
